requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-log==0.4.0
cloudinary==1.30.0
cloudinary-cli==1.5.1
docstring-parser==0.15
Flask==2.2.2
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
psycopg2-binary==2.9.5
Pygments==2.13.0
python-dotenv==0.21.0
requests==2.28.1
six==1.16.0
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.10.0
